feat: load configuration from .yml file instead of .env. fixes #5 patch: proxy related bug

+++ app/events.py
@@ -1,5 +1,6 @@
 """Startup and shutdown events"""
 
+import sys
 from datetime import timedelta
 from shutil import rmtree
 
@@ -11,39 +12,40 @@ from app.db import VideoInfo, create_tables, engine
 from app.utils import DOWNLOAD_DIR, create_temp_dirs, logger, utc_now
 
 
-def event_startup_create_tempdirs():
+def event_create_tempdirs():
     create_temp_dirs()
 
 
-def event_startup_create_tables():
+def event_create_tables():
     create_tables()
 
 
-def event_all_delete_expired_extracted_info():
+def event_delete_expired_extracted_info():
     time_offset = utc_now() - timedelta(
         hours=loaded_config.video_info_cache_period_in_hrs
     )
     delete_query = delete(VideoInfo).where(VideoInfo.updated_on < time_offset)
     with Session(bind=engine) as session:
         logger.info(f"Deleting expired extracted-infos [< {time_offset}]")
+
         session.exec(delete_query)
         session.commit()
         return time_offset
 
 
-def event_shutdown_clear_previous_downloads():
-    rmtree(DOWNLOAD_DIR)
+def event_clear_previous_downloads():
+    if loaded_config.clear_downloaded_contents:
+        if DOWNLOAD_DIR.exists():
+            print(
+                f"CLEARING download DIRECTORY {DOWNLOAD_DIR.absolute()!r} "
+                f"{loaded_config.clear_downloaded_contents=}",
+                file=sys.stderr,
+            )
+            rmtree(DOWNLOAD_DIR)
 
-
-def register_events(app: FastAPI) -> FastAPI:
-    """Sets up event handlers"""
-    for var in globals().values():
-        if callable(var):
-            if var.__name__.startswith("event_startup_"):
-                app.add_event_handler("startup", var)
-            elif var.__name__.startswith("event_shutdown_"):
-                app.add_event_handler("shutdown", var)
-            elif var.__name__.startswith("event_all_"):
-                app.add_event_handler("startup", var)
-                app.add_event_handler("shutdown", var)
-    return app
+    else:
+        print(
+            f"NOT CLEARING download DIRECTORY {DOWNLOAD_DIR.absolute()!r}. "
+            f"{loaded_config.clear_downloaded_contents=}",
+            file=sys.stderr,
+        )
